classe camera multifocal

+++ packnet_sfm/geometry/camera_multifocal_valeo.py
@@ -221,7 +221,7 @@ class CameraMultifocal(nn.Module):
               * self.scale_factors[mask, 1].unsqueeze(1).unsqueeze(2).repeat([1, H, W])).unsqueeze(1)
 
         N = 12
-        theta_tensor = (torch.zeros(B, 1, H, W)).to(device)
+        theta_tensor = (torch.zeros(B, 1, H, W)).float().to(device)
         ri = torch.sqrt(xi.pow(2) + yi.pow(2))
         for _ in range(N):
             t1 = theta_tensor
